feat: introduce smarter computer engine for moves
The computer tracks what it has played, and plays smarter moves related
to the direction a ship is currently going. One limitation this engine
has is that it doesn't work properly for ships next to each other, since
it will forget about the other ship once the first is sunk, later having
to discover the same ship, again at risk of that happening again.